refactor cont
