+++ lib/index.js
@@ -88,7 +88,7 @@ async function run() {
 
   const fullReport = await fullReportPromise;
   const numFailed = await reporter(fullReport);
-  await hookManager.runHooks('shutdown');
+  await hookManager.runHooks('Shutdown');
   if (numFailed > 0) process.exit(-1);
   else process.exit();
 }
