+++ lib/runner.js
@@ -1,26 +1,22 @@
-const Scenario = require('../models/scenario');
 const HookManager = require('../models/hook-manager');
 const {fork} = require('child_process');
-const SeleniumDriver = require('../models/selenium-driver');
+const SeleniumDriver = require('../models/selenium-interpreter');
 const driver = new SeleniumDriver();
 let childDriver = null;
 
 // interrupt graceful failure
-process.on('SIGINT', function () {
-  exitProcess();
-});
-
-process.on('SIGTERM', function () {
-  console.log(`${process.pid} - exiting...`);
-  process.exit();
-});
-
-process
+process.on('SIGINT', () => {
+    exitProcess();
+  })
+  .on('SIGTERM', () => {
+    console.log(`${process.pid} - exiting...`);
+    process.exit();
+  })
   .on('unhandledRejection', (reason, p) => {
     console.error(reason, 'Unhandled Rejection at Promise...but now I handled it', p);
     exitProcess();
   })
-  .on('uncaughtException', function (err) {
+  .on('uncaughtException', (err) => {
     console.log("Uncaught Exception...but now I caught it");
     console.log(err);
     exitProcess();
@@ -44,14 +40,6 @@ async function exitProcess(report=[]) {
       process.exit();
   }
 
-};
-
-function JSONError(err) {
-  if (!err) return 'END: ';
-  return `END: ${JSON.stringify({
-    stack: err.stack,
-    message: err.message
-  })}`;
 }
 
 async function executeScenarios(scenarios) {
@@ -81,29 +69,7 @@ async function executeScenarios(scenarios) {
           if (regexResults[1] !== '') resolve(JSON.parse(regexResults[1]));
           else resolve();
         }
-        else if (/WD: (\d+) - (.*)/.test(m)) {
-          const regexResults = m.match(/WD: (\d+) - (.*)/);
-          const uniqueId = regexResults[1];
-          console.log('WD MESSAGE RECEIVED', m);
-
-          driver.callWDFunction(regexResults[2]).then((result) => {
-            if (childDriver) childDriver.send(`WD: ${uniqueId} - ${result}`);
-          }).catch((err) => {
-            if (childDriver) childDriver.send(JSONError(err));
-          });
-        }
-        else if (/WDE: (\d+) - (.*)/.test(m)) {
-          const regexResults = m.match(/WDE: (\d+) - (.*)/);
-          const uniqueId = regexResults[1];
-          console.log('WDE MESSAGE RECEIVED', m);
-
-          driver.callWDElementFunction(regexResults[2]).then((result) => {
-            if (childDriver) childDriver.send(`WDE: ${uniqueId} - ${result}`);
-          }).catch((err) => {
-            if (childDriver) childDriver.send(JSONError(err));
-          });
-        }
-
+        driver.interpret(m, childDriver);
       });
     });
